+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+* Add basic support for multiple inputs in Tectonic.toml
 * Improve performance of file set cache used by inspections
 * Support label references to user defined listings environment
 * Add option to disable automatic compilation in power save mode